Upgrade: datasetManager, networkSolver, dataTypes, helper
datasetManager:
- renamed methods: extractRealSamples, createSynthSamples, generateAndStoreSamples
- computes maxSimTmpl instead of the solver
- removed unnecessary methods
- generateAndStoreSamples detects automatically if samples need to be created
- removed quaternion vectors

dataTypes:
- Sample struct now has getQuat() and getTrans() methods to return a quaternion or a translation vector
- loadbar, showRGBDPatch, to_array functions are moved to helper.h/cpp

helper (contains helper functions):
-  loadbar, showRGBDPatch, to_array, fexists, depth2normals

hdf5handler:
- modified readBBIntristicMats for BigBIRD dataset

networkSolver:
- computeMaxSimTmpl moved to datasetManager
- threaded batch creation is activated by a preprocessor directive

+++ src/networkevaluator.cpp
@@ -95,7 +95,7 @@ void networkEvaluator::visualizeKNN(caffe::Net<float> &CNN,
     Mat DBfeats, DBtest;
     std::random_device ran;
 
-    for (int obj = 0; obj < test_set.size(); ++obj) {
+    for (size_t obj = 0; obj < test_set.size(); ++obj) {
         DBtest.push_back(computeDescriptors(CNN, test_set[obj]));
         DBfeats.push_back(computeDescriptors(CNN, templates[obj]));
     }
